Add login route & component
Every route that does not have the "noAuth" meta flag gets redirected to the login page. After successfully logging in, the user gets redirected  back to the originally called route.

See Redmine ticket #11203.
